servers/v2t/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-generativeai>=0.8.0
gunicorn==21.2.0
